Some more test fixes.
