+++ tests/ManagerTest.php
@@ -62,7 +62,6 @@ class ManagerTest extends AbstractTestCase
         $this->blacklist = Mockery::mock(Blacklist::class);
         $this->factory = Mockery::mock(Factory::class);
         $this->manager = new Manager($this->jwt, $this->blacklist, $this->factory);
-
         $this->validator = Mockery::mock(PayloadValidator::class);
     }
