+++ tests/Middleware/AbstractMiddlewareTest.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Tymon\JWTAuth\Test\Middleware;
+
+use Mockery;
+use Tymon\JWTAuth\JWTAuth;
+use Illuminate\Http\Request;
+use Tymon\JWTAuth\Test\AbstractTestCase;
+
+abstract class AbstractMiddlewareTest extends AbstractTestCase
+{
+    /**
+     * @var \Mockery\MockInterface|\Tymon\JWTAuth\JWTAuth
+     */
+    protected $auth;
+
+    /**
+     * @var \Mockery\MockInterface|\Illuminate\Http\Request
+     */
+    protected $request;
+
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->auth = Mockery::mock(JWTAuth::class);
+        $this->request = Mockery::mock(Request::class);
+    }
+
+    public function tearDown()
+    {
+        Mockery::close();
+
+        parent::tearDown();
+    }
+}
