+++ tests/Middleware/AuthenticateAndRenewTest.php
@@ -12,27 +12,14 @@
 namespace Tymon\JWTAuth\Test\Middleware;
 
 use Mockery;
-use Tymon\JWTAuth\JWTAuth;
-use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Tymon\JWTAuth\Http\Parser\Parser;
 use Tymon\JWTAuth\Test\Stubs\UserStub;
-use Tymon\JWTAuth\Test\AbstractTestCase;
 use Tymon\JWTAuth\Exceptions\TokenInvalidException;
 use Tymon\JWTAuth\Http\Middleware\AuthenticateAndRenew;
 
-class AuthenticateAndRenewTest extends AbstractTestCase
+class AuthenticateAndRenewTest extends AbstractMiddlewareTest
 {
-    /**
-     * @var \Mockery\MockInterface|\Tymon\JWTAuth\JWTAuth
-     */
-    protected $auth;
-
-    /**
-     * @var \Mockery\MockInterface
-     */
-    protected $request;
-
     /**
      * @var \Tymon\JWTAuth\Http\Middleware\Authenticate|\Tymon\JWTAuth\Http\Middleware\AuthenticateAndRenew
      */
@@ -42,17 +29,7 @@ class AuthenticateAndRenewTest extends AbstractTestCase
     {
         parent::setUp();
 
-        $this->auth = Mockery::mock(JWTAuth::class);
-        $this->request = Mockery::mock(Request::class);
-
         $this->middleware = new AuthenticateAndRenew($this->auth);
-    }
-
-    public function tearDown()
-    {
-        Mockery::close();
-
-        parent::tearDown();
     }
 
     /** @test */
