+++ tests/Validators/PayloadValidatorTest.php
@@ -32,7 +32,7 @@ class PayloadValidatorTest extends AbstractTestCase
     {
         parent::setUp();
 
-        $this->validator = new PayloadValidator();
+        $this->validator = new PayloadValidator;
     }
 
     /** @test */
